tests/requirements.txt
connexion[swagger-ui] == 2.3.0
flask-sqlalchemy == 2.4.0
flask-migrate == 2.5.2
psycopg2-binary==2.7.7